healthbackend/pyproject.toml
[tool.poetry]
name = "healthbackend"
version = "0.1.0"
description = "health manage"
authors = ["wangyunshan"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.16"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-rest-swagger = "^2.2.0"
django-filter = "^24.3"
Faker = "^33.0.0"
numpy = "^2.1.3"
drf-yasg = "^1.21.8"
django-cors-headers = "^4.6.0"
pillow = "^11.0.0"
django-adminlte-ui = "^2.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
package-mode = false

[[tool.poetry.source]]
name = "mirrors"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"